updating to run from guide tree and representative alignment stage
