+++ environment.yml
@@ -8,11 +8,11 @@ dependencies:
   - iqtree
   - mafft
   - minimap2=2.17
+  - iqtree
   - pip=19.3.1
   - python=3.6
   - snakemake-minimal=5.13
   - pip:
-    - git+https://github.com/Cibiv/IQ-TREE.git@v2.0-rc2
-    - git+https://github.com/cov-ert/datafunk.git@v0.0.3
+    # - git+https://github.com/Cibiv/IQ-TREE.git@v2.0-rc2
     - git+https://github.com/cov-ert/clusterfunk.git@v0.0.1
     - pytools==2020.1
