+++ lineage/__init__.py
@@ -1,0 +1,2 @@
+_program = "lineage"
+__version__ = "0.1.0"
