+++ lineage/config.yaml
@@ -1,0 +1,4 @@
+lineage_metadata: 'data/lineages.header_fixed.csv'
+fasta: 'data/GISAID_2020-03-31_2685_CDS_95%.header_fixed.fasta'
+representative_aln: "data/representative_sequences.aln.iqtree_friendly.fasta"
+guide_tree: "data/representative_sequences.aln.iqtree_friendly.fasta.treefile"
