+++ setup.py
@@ -4,15 +4,15 @@ import os
 import pkg_resources
 # Note: the _program variable is set in __init__.py.
 # it determines the name of the package/final command line tool.
-from subtype import __version__, _program
+from lineage import __version__, _program
 
-setup(name='subtype',
+setup(name='lineage',
       version=__version__,
-      packages=['subtype'],
-      scripts=['subtype/bin/all_lineage_assignment.smk',
-                'subtype/bin/assign_query_lineage.smk',
-                'subtype/bin/Snakefile'],
-      package_data={'subtype':['config.yaml',
+      packages=['lineage'],
+      scripts=['lineage/bin/assign_query_file.smk',
+                'lineage/bin/assign_query_lineage.smk',
+                'lineage/bin/Snakefile'],
+      package_data={'lineage':['config.yaml',
                                 'data/*']},
       description='hcov-2019 subtyping command line tool',
       url='',
@@ -20,7 +20,7 @@ setup(name='subtype',
       author_email='aine.otoole@ed.ac.uk',
       entry_points="""
       [console_scripts]
-      {program} = subtype.command:main
+      {program} = lineage.command:main
       """.format(program = _program),
       include_package_data=True,
       keywords=[],
