--- subtype/__init__.py
@@ -1,2 +1,0 @@
-_program = "subtype"
-__version__ = "0.1.0"
