--- subtype/bin/Snakefile
@@ -1,29 +1,0 @@
-configfile: workflow.current_basedir + "/../config.yaml"
-
-import csv
-from Bio import SeqIO
-from pytools.persistent_dict import PersistentDict
-
-##### Configuration #####
-
-if config.get("lineage_metadata"):
-    config["lineage_metadata"] = os.path.join(workflow.current_basedir,'..', config["lineage_metadata"])
-if config.get("fasta"):
-    config["fasta"] = os.path.join(workflow.current_basedir,'..', config["fasta"])
-
-if config.get("outdir"):
-    config["outdir"] = config["outdir"].rstrip("/")
-else:
-    config["outdir"] = "analysis"
-
-##### Target rules #####
-
-rule all:
-    input:
-        config["outdir"] + "/temp/representative_sequences.aln.fasta",
-        config["outdir"] + "/temp/representative_sequences.aln.fasta.treefile",
-        config["outdir"] + "/temp/temp_report.txt",
-        config["outdir"] + "/lineage_report.txt"
-
-##### Modules #####
-include: "all_lineage_assignment.smk"
