--- subtype/config.yaml
@@ -1,2 +1,0 @@
-lineage_metadata: 'data/lineages.header_fixed.csv'
-fasta: 'data/GISAID_2020-03-31_2685_CDS_95%.header_fixed.fasta'
